--- Ontologies/wearableprivacyvuln.ttl
@@ -1,100 +1,0 @@
-@prefix : <https://emikr15.student,sdu.dk/21/10/05/wearableprivacyvunl.ttl> .
-@prefix owl: <http://www.w3.org/2002/07/owl#> .
-@prefix rdf: <http://www.w3.org/1999/02/22-rdf-syntax-ns#> .
-@prefix xml: <http://www.w3.org/XML/1998/namespace> .
-@prefix xsd: <http://www.w3.org/2001/XMLSchema#> .
-@prefix rdfs: <http://www.w3.org/2000/01/rdf-schema#> .
-@prefix privacyvunl: <https://ontology.hviidnet.com/2020/01/03/privacyvunl.ttl#> .
-@prefix privacyvunlv2: <https://ontology.hviidnet.com/2020/01/03/privacyvunlV2.ttl#> .
-
-<https://emikr15.student,sdu.dk/21/10/05/wearableprivacyvunl.ttl#> rdf:type owl:Ontology ;
-	owl:imports <https://ontology.hviidnet.com/2020/01/03/privacyvunlV2.ttl#> .
-
-# contexts
-# e.g. smart devices
-
-
-:BodyPart a owl:class ;
-    owl:disjointWith :Environment ;
-    rdfs:subClassOf privacyvunlv2:Context .
-
-:Arm a owl:class ;
-    rdfs:subClassOf privacyvunlv2:Context .
-
-:UpperArm a owl:class ;
-    rdfs:subClassOf privacyvunlv2:Context .
-
-:Wrist a owl:class ;
-    rdfs:subClassOf privacyvunlv2:Context .
-
-:Head a owl:class ;
-    rdfs:subClassOf privacyvunlv2:Context .
-
-:Leg a owl:class ;
-    rdfs:subClassOf privacyvunlv2:Context .
-
-:UpperLeg a owl:class ;
-    rdfs:subClassOf privacyvunlv2:Context .
-
-:LowerLeg a owl:class ;
-    rdfs:subClassOf privacyvunlv2:Context .
-
-:Foot a owl:class ;
-    rdfs:subClassOf privacyvunlv2:Context .
-
-:Back a owl:class ;
-    rdfs:subClassOf privacyvunlv2:Context .
-
-:Chesta owl:class ;
-    rdfs:subClassOf privacyvunlv2:Context .
-
-:Environment a owl:class ;
-    owl:disjointWith :BodyPart ;
-    rdfs:subClassOf privacyvunlv2:Context .
-
-
-# data
-
-:Accelerometer a owl:NamedIndividual ;
-    a privacyvunl:TimeSeries .
-
-:ElectrodermalActivity a owl:NamedIndividual ;
-    a privacyvunl:TimeSeries ;
-    owl:equivalentClass :GalvanicSkinResponse .
-
-:GalvanicSkinResponse a owl:NamedIndividual ;
-    a privacyvunl:TimeSeries ;
-    owl:equivalentClass :ElectrodermalActivity .
-
-:Photoplethysmography a owl:NamedIndividual ;
-    a privacyvunl:TimeSeries .
-
-:Electrocardiography a owl:NamedIndividual ;
-    a privacyvunl:TimeSeries .
-
-
-:Thermometer a owl:NamedIndividual ;
-    a privacyvunl:TimeSeries .
-
-# Transformations
-
-:InterbeatInterval a owl:NamedIndividual ;
-    a privacyvunl:TimeSeries .
-
-:PhysicalActivity a owl:NamedIndividual ;
-    a privacyvunl:TimeSeries .
-
-:SleepPattern a owl:NamedIndividual ;
-    a privacyvunl:TimeSeries .
-
-:BloodVolumePulse a owl:NamedIndividual ;
-    a privacyvunl:TimeSeries .
-
-:BreathingRate a owl:NamedIndividual ;
-    a privacyvunl:TimeSeries .
-
-:HeartRate a owl:NamedIndividual ;
-    a privacyvunl:TimeSeries .
-
-:HandGestures a owl:NamedIndividual ;
-    a privacyvunl:TimeSeries .
