+++ Templates/wearableprivacyvunl/PrivacyAttacks/Gender.py
@@ -1,13 +1,14 @@
-# https://ieeexplore-ieee-org.proxy1-bib.sdu.dk/document/7917140
-from rdflib import Namespace, Literal, XSD, RDF
+from rdflib import Namespace, RDF, XSD
+from rdflib.term import Literal
 
 from ITemplate import IPrivacyAttack
 
 
-class KeyboardInput(IPrivacyAttack):
+# https://ieeexplore-ieee-org.proxy1-bib.sdu.dk/document/7514649
+class Gender(IPrivacyAttack):
     # noinspection SpellCheckingInspection
     __DOMAINNAMESPACE__: Namespace = Namespace(
-        "https://emikr15.student,sdu.dk/21/10/05/wearableprivacyvunl.ttl#"
+        "https://emikr15.student.sdu.dk/21/10/05/wearableprivacyvunl.ttl#"
     )
 
     def __init__(self):
@@ -20,23 +21,23 @@ class KeyboardInput(IPrivacyAttack):
         accelerometer = self.MODELS["inputRequirementACC"]
         triples = [
             (accelerometer, self.RDF.type, self.PRIVVULNV2.Constraint),
-            (accelerometer, self.PRIVVULNV2.TemporalResolution, Literal("0.02", datatype=self.XSD.double)),
+            (accelerometer, self.PRIVVULNV2.TemporalResolution, Literal("0.01", datatype=self.XSD.double)),
             (accelerometer, self.PRIVVULN.feeds, self.__DOMAINNAMESPACE__.PhysicalActivity)
         ]
 
-        transformation = self.MODELS["accToKeyboardInput"]
+        transformation = self.MODELS["accelerometerToGender"]
         triples += [
             (transformation, RDF.type, self.PRIVVULN.PrivacyAttack),
             (accelerometer, self.PRIVVULN.feeds, transformation),
         ]
 
-        keyboard_input = self.MODELS["KeyboardInput"]
+        gender = self.MODELS["Gender"]
         triples += [
-            (keyboard_input, RDF.type, self.PRIVVULN.PrivacyRisk),
-            (keyboard_input, self.PRIVVULNV2.description, Literal("This is bad!", datatype=XSD.string)),
-            (transformation, self.PRIVVULN.creates, keyboard_input),
+            (gender, RDF.type, self.PRIVVULN.PrivacyRisk),
+            (gender, self.PRIVVULNV2.description, Literal("This is bad!", datatype=XSD.string)),
+            (transformation, self.PRIVVULN.creates, gender),
             # TODO fix score
-            (keyboard_input, self.PRIVVULNV2.privacyRiskScore, Literal(5, datatype=XSD.int))
+            (gender, self.PRIVVULNV2.privacyRiskScore, Literal(5, datatype=XSD.int))
         ]
 
         [self.graph.add(triple) for triple in triples]
