+++ Templates/wearableprivacyvunl/Transformations/HeartRate.py
@@ -7,7 +7,7 @@ from Templates.ITemplate import ITransformation
 class HeartRate(ITransformation):
     # noinspection SpellCheckingInspection
     __DOMAINNAMESPACE__ = Namespace(
-        "https://emikr15.student,sdu.dk/21/10/05/wearableprivacyvunl.ttl#"
+        "https://emikr15.student.sdu.dk/21/10/05/wearableprivacyvunl.ttl#"
     )
 
     def __init__(self):
@@ -30,36 +30,36 @@ class HeartRate(ITransformation):
             (input_node_ecg, self.PRIVVULN.feeds, time_resolution_ecg)
         ]
 
-        input_node_ppg = self.MODELS["inputRequirementPPG"]
-        triples = [
-            (input_node_ppg, self.RDF.type, self.PRIVVULNV2.Constraint),
-            (input_node_ppg, self.PRIVVULN.feeds, self.__DOMAINNAMESPACE__.Photoplethysmography)
-        ]
-        time_resolution_ppg = self.MODELS['timeResolutionLinearPPG']
-        triples += [
-            (time_resolution_ppg, RDF.type, self.PRIVVULNV2.TimeResolutionLinear),
-            (time_resolution_ppg, self.PRIVVULNV2.TimeInput, Literal(1.0, datatype=self.XSD.double)),
-            (time_resolution_ppg, self.PRIVVULNV2.TimeOutput, Literal(1.0, datatype=self.XSD.double)),
-            (input_node_ppg, self.PRIVVULN.feeds, time_resolution_ppg)
-        ]
+        # input_node_ppg = self.MODELS["inputRequirementPPG"]
+        # triples = [
+        #     (input_node_ppg, self.RDF.type, self.PRIVVULNV2.Constraint),
+        #     (input_node_ppg, self.PRIVVULN.feeds, self.__DOMAINNAMESPACE__.Photoplethysmography)
+        # ]
+        # time_resolution_ppg = self.MODELS['timeResolutionLinearPPG']
+        # triples += [
+        #     (time_resolution_ppg, RDF.type, self.PRIVVULNV2.TimeResolutionLinear),
+        #     (time_resolution_ppg, self.PRIVVULNV2.TimeInput, Literal(1.0, datatype=self.XSD.double)),
+        #     (time_resolution_ppg, self.PRIVVULNV2.TimeOutput, Literal(1.0, datatype=self.XSD.double)),
+        #     (input_node_ppg, self.PRIVVULN.feeds, time_resolution_ppg)
+        # ]
 
-        ecg_to_hr = self.MODELS["accelerometerToPhysicalActivity"]
+        ecg_to_hr = self.MODELS["ecgToHeartRate"]
         triples += [
             (ecg_to_hr, RDF.type, self.PRIVVULN.Transformation),
             (input_node_ecg, self.PRIVVULN["feeds"], ecg_to_hr)
         ]
 
-        ppg_to_hr = self.MODELS["accelerometerToPhysicalActivity"]
-        triples += [
-            (ppg_to_hr, RDF.type, self.PRIVVULN.Transformation),
-            (input_node_ecg, self.PRIVVULN["feeds"], ppg_to_hr)
-        ]
+        # ppg_to_hr = self.MODELS["ppgToHeartRate"]
+        # triples += [
+        #     (ppg_to_hr, RDF.type, self.PRIVVULN.Transformation),
+        #     (input_node_ecg, self.PRIVVULN["feeds"], ppg_to_hr)
+        # ]
 
         physical_activity = self.MODELS["InterbeatInterval"]
         triples += [
             (physical_activity, RDF.type, self.PRIVVULN.TimeSeries),
             (physical_activity, RDF.type, self.__DOMAINNAMESPACE__.InterbeatInterval),
-            (ppg_to_hr, self.PRIVVULN.feeds, physical_activity),
+            # (ppg_to_hr, self.PRIVVULN.feeds, physical_activity),
             (ecg_to_hr, self.PRIVVULN.feeds, physical_activity)
 
         ]
