+++ Templates/wearableprivacyvunl/Transformations/SleepPattern.py
@@ -7,7 +7,7 @@ from ITemplate import ITransformation
 # https://dl.acm.org/doi/pdf/10.1145/3309074.3309076
 class SleepPattern(ITransformation):
     __DOMAINNAMESPACE__ = Namespace(
-        "https://emikr15.student,sdu.dk/21/10/05/wearableprivacyvunl.ttl#"
+        "https://emikr15.student.sdu.dk/21/10/05/wearableprivacyvunl.ttl#"
     )
 
     def __init__(self):
