+++ Templates/wearableprivacyvunl/Transformations/__init__.py
@@ -1,0 +1,5 @@
+import glob
+from os.path import dirname, basename, isfile, join
+
+modules = glob.glob(join(dirname(__file__), "*.py"))
+__all__ = [basename(f)[:-3] for f in modules if isfile(f) and not f.endswith('__init__.py')]
